# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rippl - Patient Referral Reward Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Referrer {
  id: string;
  patient_id: string;
  name: string;
  phone: string;
  email?: string;
  referral_code: string;
  total_referrals: number;
  total_rewards_issued: number;
  created_at: string;
}

export interface CreateReferrerInput {
  patient_id: string;
  name: string;
  phone: string;
  email?: string;
}

export interface ReferrerQr {
  referral_url: string;
  referral_code: string;
}

export type ReferralEventTeamSource =
  (typeof ReferralEventTeamSource)[keyof typeof ReferralEventTeamSource];

export const ReferralEventTeamSource = {
  front: "front",
  back: "back",
  assistant: "assistant",
} as const;

export type ReferralEventStatus =
  (typeof ReferralEventStatus)[keyof typeof ReferralEventStatus];

export const ReferralEventStatus = {
  Lead: "Lead",
  Booked: "Booked",
  Exam_Completed: "Exam Completed",
  Reward_Sent: "Reward Sent",
} as const;

export interface ReferralEvent {
  id: string;
  new_patient_name: string;
  new_patient_phone: string;
  referrer_id: string;
  referrer_name?: string;
  team_source: ReferralEventTeamSource;
  office: string;
  status: ReferralEventStatus;
  reward_type?: string | null;
  created_at: string;
}

export type CreateReferralInputTeamSource =
  (typeof CreateReferralInputTeamSource)[keyof typeof CreateReferralInputTeamSource];

export const CreateReferralInputTeamSource = {
  front: "front",
  back: "back",
  assistant: "assistant",
} as const;

export interface CreateReferralInput {
  new_patient_name: string;
  new_patient_phone: string;
  referrer_id: string;
  team_source: CreateReferralInputTeamSource;
  office: string;
}

export type UpdateReferralStatusInputStatus =
  (typeof UpdateReferralStatusInputStatus)[keyof typeof UpdateReferralStatusInputStatus];

export const UpdateReferralStatusInputStatus = {
  Lead: "Lead",
  Booked: "Booked",
  Exam_Completed: "Exam Completed",
  Reward_Sent: "Reward Sent",
} as const;

export interface UpdateReferralStatusInput {
  status: UpdateReferralStatusInputStatus;
}

export interface ReferralWithReferrer {
  referral: ReferralEvent;
  referrer: Referrer;
}

export type RewardRewardType =
  (typeof RewardRewardType)[keyof typeof RewardRewardType];

export const RewardRewardType = {
  "in-house-credit": "in-house-credit",
  "amazon-gift-card": "amazon-gift-card",
  "partner-gift-card": "partner-gift-card",
} as const;

export interface Reward {
  id: string;
  referrer_id: string;
  referral_event_id: string;
  reward_type: RewardRewardType;
  fulfilled: boolean;
  created_at: string;
}

export type CreateRewardInputRewardType =
  (typeof CreateRewardInputRewardType)[keyof typeof CreateRewardInputRewardType];

export const CreateRewardInputRewardType = {
  "in-house-credit": "in-house-credit",
  "amazon-gift-card": "amazon-gift-card",
  "partner-gift-card": "partner-gift-card",
} as const;

export interface CreateRewardInput {
  referrer_id: string;
  referral_event_id: string;
  reward_type: CreateRewardInputRewardType;
}

export interface TopReferrer {
  id: string;
  name: string;
  total_referrals: number;
  total_rewards_issued: number;
}

export interface DashboardStats {
  total_referrals: number;
  exams_completed: number;
  rewards_issued: number;
  active_referrers: number;
  top_referrers: TopReferrer[];
  recent_events: ReferralEvent[];
}