# lib/api-zod/src/generated/types/createReferralInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rippl - Patient Referral Reward Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateReferralInputTeamSource } from "./createReferralInputTeamSource";

export interface CreateReferralInput {
  new_patient_name: string;
  new_patient_phone: string;
  referrer_id: string;
  team_source: CreateReferralInputTeamSource;
  office: string;
}