# lib/api-zod/src/generated/types/createRewardInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rippl - Patient Referral Reward Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateRewardInputRewardType } from "./createRewardInputRewardType";

export interface CreateRewardInput {
  referrer_id: string;
  referral_event_id: string;
  reward_type: CreateRewardInputRewardType;
}