# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rippl - Patient Referral Reward Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ReferralEvent } from "./referralEvent";
import type { TopReferrer } from "./topReferrer";

export interface DashboardStats {
  total_referrals: number;
  exams_completed: number;
  rewards_issued: number;
  active_referrers: number;
  top_referrers: TopReferrer[];
  recent_events: ReferralEvent[];
}