# lib/api-zod/src/generated/types/referralEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rippl - Patient Referral Reward Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ReferralEventStatus } from "./referralEventStatus";
import type { ReferralEventTeamSource } from "./referralEventTeamSource";

export interface ReferralEvent {
  id: string;
  new_patient_name: string;
  new_patient_phone: string;
  referrer_id: string;
  referrer_name?: string;
  team_source: ReferralEventTeamSource;
  office: string;
  status: ReferralEventStatus;
  reward_type?: string | null;
  created_at: string;
}