# lib/api-zod/src/generated/types/referralWithReferrer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rippl - Patient Referral Reward Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ReferralEvent } from "./referralEvent";
import type { Referrer } from "./referrer";

export interface ReferralWithReferrer {
  referral: ReferralEvent;
  referrer: Referrer;
}