# lib/api-zod/src/generated/types/referrer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rippl - Patient Referral Reward Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Referrer {
  id: string;
  patient_id: string;
  name: string;
  phone: string;
  email?: string;
  referral_code: string;
  total_referrals: number;
  total_rewards_issued: number;
  created_at: string;
}