# lib/api-zod/src/generated/types/topReferrer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rippl - Patient Referral Reward Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface TopReferrer {
  id: string;
  name: string;
  total_referrals: number;
  total_rewards_issued: number;
}