# lib/api-zod/src/generated/types/updateReferralStatusInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rippl - Patient Referral Reward Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateReferralStatusInputStatus } from "./updateReferralStatusInputStatus";

export interface UpdateReferralStatusInput {
  status: UpdateReferralStatusInputStatus;
}